separated knight.cpp into a header and the corresponding file.
